Removed ui related code from shared directory
