+++ ui/ducks/send/helpers.js
@@ -1,6 +1,6 @@
 import { addHexPrefix } from 'ethereumjs-util';
 import abi from 'human-standard-token-abi';
-import { GAS_LIMITS } from '../../../shared/constants/gas';
+import { GAS_LIMITS, MIN_GAS_LIMIT_HEX } from '../../../shared/constants/gas';
 import { CHAIN_ID_TO_GAS_LIMIT_BUFFER_MAP } from '../../../shared/constants/network';
 import {
   ASSET_TYPES,
@@ -13,7 +13,6 @@ import {
 } from '../../../shared/modules/conversion.utils';
 import { ETH, GWEI } from '../../helpers/constants/common';
 import { calcTokenAmount } from '../../helpers/utils/token-util';
-import { MIN_GAS_LIMIT_HEX } from '../../pages/send/send.constants';
 import {
   addGasBuffer,
   generateERC20TransferData,
