+++ ui/pages/send/send.constants.js
@@ -3,12 +3,12 @@ import {
   multiplyCurrencies,
 } from '../../../shared/modules/conversion.utils';
 import { addHexPrefix } from '../../../app/scripts/lib/util';
+import { MIN_GAS_LIMIT_HEX } from '../../../shared/constants/gas';
 
 const MIN_GAS_PRICE_DEC = '0';
 const MIN_GAS_PRICE_HEX = parseInt(MIN_GAS_PRICE_DEC, 10).toString(16);
 const MIN_GAS_LIMIT_DEC = '21000';
 const MAX_GAS_LIMIT_DEC = '7920027';
-const MIN_GAS_LIMIT_HEX = parseInt(MIN_GAS_LIMIT_DEC, 10).toString(16);
 
 const HIGH_FEE_WARNING_MULTIPLIER = 1.5;
 const MIN_GAS_PRICE_GWEI = addHexPrefix(
@@ -63,7 +63,6 @@ export {
   ENS_UNKNOWN_ERROR,
   ENS_REGISTRATION_ERROR,
   MIN_GAS_LIMIT_DEC,
-  MIN_GAS_LIMIT_HEX,
   MIN_GAS_PRICE_DEC,
   MIN_GAS_PRICE_GWEI,
   MIN_GAS_PRICE_HEX,
