+++ ui/pages/send/send.utils.js
@@ -10,7 +10,7 @@ import {
 
 import { calcTokenAmount } from '../../helpers/utils/token-util';
 import { addHexPrefix } from '../../../app/scripts/lib/util';
-import { ERC20, ERC721 } from '../../helpers/constants/common';
+import { ERC20, ERC721 } from '../../../shared/constants/transaction';
 import {
   TOKEN_TRANSFER_FUNCTION_SIGNATURE,
   COLLECTIBLE_TRANSFER_FROM_FUNCTION_SIGNATURE,
